requirements.txt
asgiref==3.3.4
chronograph==1.0.0
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
future==0.18.2
gunicorn==20.1.0
iso8601==0.1.16
livereload==2.6.3
Pillow==8.2.0
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1
PyYAML==5.4.1
schedule==1.1.0
serial==0.0.97
six==1.16.0
sqlparse==0.4.1
tornado==6.1
whitenoise==5.3.0
